Check SOAP fault fields as implicit receive conditions
Checking for whether a SOAP fault was expected or not shouldn't be performed
during the deconstruction of the incoming message. In fact, it doesn't really
have anything to do with it: we would still like to deconstruct the incoming
message even when it's the wrong kind of SOAP message.

Additionally, checks for the fault code or the fault string should be performed
in the same place.

This commit moves all those checks to the
ReceiveDataSpecification#validateConditions method, modeling them as implicit
error conditions.

The literal data for a SOAP fault message is no longer only its details, but rather
the whole contents of the soap:Body element, as in regular messages. This is more
consistent and allows the user to specify more fine-grained conditions on the fault
code, string, or actors. These weren't accesible by the XPath queries before.

From: Antonio García <antonio.garciadominguez@uca.es>