IBPELDeployer: added cleanUpAfterTestCase()
Later testing showed that just killing all processes after undeploying the test
case didn't work for long enough test suites. Basically, soon as we hit on as
many defective WS-BPEL compositions as the number of CPUs in our machine, we'll
be out of luck again. Lingering WS-BPEL processes need to be killed after every
test case.

This commit only adds the cleanUpAfterTestCase method to the IBPELDeployer
interface and all classes which implement it, and calls it after every test
case. The method itself is just a placeholder for now: it will be implemented
for ActiveBPEL later.

From: Antonio García <antonio.garciadominguez@uca.es>