WSDLParser: add support for rpc/lit messages
This commit allows WSDLParser to generate sample content for RPC/literal
operations, adding support for rpc/lit operations in the graphical BPTS
editor. There's one catch, though: BPELUnit assumed that all literal data
for RPC operations is not wrapped, but the editor produces wrapped content.

However, changing the editor to produce unwrapped data for RPC operations in
the BPTS would be quite difficult the way it's written now: half the code there
depends on having a single XML element to place in the BPTS.

The best thing would be to force all literal data for RPC-style operations to
be wrapped in the BPTS, but that'd break backwards compatibility.

Instead, we handle both cases. The wrapper detection is not 100% accurate: it
will probably get confused in some corner cases. For example, it could be a
problem if the first part in the message had the same name as the operation.

It is still undecided if the old unwrapped format will be deprecated or not.

From: Antonio García Domínguez <antonio.garciadominguez@uca.es>