OperationDataComponent: replace "send fault" checkbox with message list
The current message editor can only produce XML snippets for the <output>
element of the WSDL operation. Even if we check the box asking for a fault
to be sent, we won't be able to really edit its contents.

This is in part due to the fact that the editor is not aware that the same
operation may produce different messages. This commit replaces the old
"send a fault" checkbox with a selector that allows the user to pick between
the message type used for regular replies and the message types used for
each of the errors. The editor assigns the value of the old "fault" attribute
automatically. There's also an entry for entering a custom fault: if selected,
no sample element will be generated.

For the time being, the name of the selected fault will be saved in the
faultstring attribute of the activity. This could be a problem if the user
wanted to send a custom description: in the future, it would be better to
have it use its own faultname attribute.

From: Antonio García Domínguez <antonio.garciadominguez@uca.es>