OperationChangeListener: renamed to MessageChangeListener
This kind of listener is notified when the user changes the WSDL operation.
However, what it's really useful for is to notify the listener that the
message to be edited has changed, and that a new XML fragment should be used.

This doesn't have to mean that the operation changed: for instance, the user
could have changed the message to be sent from a regular, successful reply to
a fault message instead.

This commit renames all the "operation change" identifiers to "message change"
identifiers, so the next commits make more sense.

From: Antonio García Domínguez <antonio.garciadominguez@uca.es>