WSDLParser: reject rpc/lit parts with the element attribute
This commit complements the previous commit which rejected doc/lit parts
with the type attribute. This is the other side of the coin: the element
attribute cannot be used on the parts of a message which is used in an
rpc/lit operation. The user should also be notified of this.

From: Antonio García Domínguez <antonio.garciadominguez@uca.es>