WSDLParser: collect schemata from imported WSDL files
The WS-I Basic Profile 1.1 says that schema elements which are only available
from a nested import should not be visible to the importing WSDL document.
However, it might be the case that we import a message which uses a type that
is defined in the imported WSDL's <types> section.

If we don't collect the schemata from the imported WSDL files, we'll end up
with an ugly NullPointerException while drawing the MessageEditor for that
message. This commit fixes that problem by taking advantage of the fact that
WSDL4J 1.6.2 kindly collects recursively all imports into the Map returned by
getImports: its own code calls readWSDL again on these imports, adding them
to the Map in the main WSDL.

From: Antonio García Domínguez <antonio.garciadominguez@uca.es>