Previously received message is now available in the template
To make smarter mockups, we needed to somehow take the received message into
account. This commit lets the user do just that: templates can now use the
$request variable, which will contain an org.w3c.dom.Element object with the
SOAP body of the received message.

Since getting anything from a DOM object is downright annoying, the user can
take advantage of a XPathTool instance in $xpath for running simple XPath 1.0
queries on it. XPathTool currently has evaluateAsList (for returning a
List<Node> with all results) and evaluateAsString, but more will be added in
the future if it is deemed necessary.

I originally wanted to use XmlObjects and use their selectPath methods, but
that'd have required adding another 5MB worth of .jars to the lib/ directory
for just a few XPath 2.0 queries which wouldn't be all that complicated most
of the time.

This commit also adds the required test case to ensure everything works as
expected.