Set up XPathTool with proper prefix/NS mappings
I need to make sure namespace prefixes work in the XPath queries used
from the Velocity templates. Every prefix which is available for the
<receive> conditions should be also available to XPathTool.

To test this, I've changed all test cases to use
elementFormDefault="qualified", so the XPath query in the test case
which uses XPathTool will need to use a namespace prefix.