SpecificationLoader: iterate over rows and then over variations
If a test case had variations in the delays in its activities, the
BPTS loader created several copies of that test case, changing the
delays in each copy. This way, TestCase did not have to know anything
about variations.

However, this doesn't work for data sources, as Velocity contexts are
created on the fly during execution of the test case, since they might
contain variables whose values is only known at that time (request, for
instance).

For that reason, TestCase got a few new getters and setters for the
IDataSource instance and the row it has to load into the contexts it
creates. The SpecificationLoader is responsible for creating and
initialising the data source and calling TestCase's setters, so TestCase
will only need to initialize the context with the row it's been told to use.

As SpecificationLoader was getting far too big, the data source initialisation
code is in a separate class in the data source package: DataSourceUtil. This
commit only adds support for inline data source contents: the following commits
will add more source reference types.