Back to random access for data sources
Data sources actually *need* random access. The SpecificationLoader, when faced
with variations, produces several copies of the same test suite, with different
delays. However, next() should not be called for them. But, the code which is
responsible for that call does not know about it, and cannot know about it:
instead, the right row index set in the TestCase by the SpecificationLoader should
be used.

This also provides for a way to optimize data source accesses in the future, by
reusing the same data source contents over several test cases and test suites. This
should be useful for resources loaded through network connections.