ActiveBPELDeployer: autodetect BPR directory when CATALINA_HOME is set
The engine-based end-to-end tests require that the deployment directory
property in conf/configuration.xml (or the .bpts) points to the right
absolute path. This is fine when we're working on our own system, but
we have no way to know where the Tomcat instance running ActiveBPEL is
in the developer's machine, and we shouldn't have to customize it to
get the tests to run.

This commit adds some logic to the ActiveBPEL deployer so it will use
$CATALINA_HOME/bpr as deployment directory if the CATALINA_HOME environment
variable is set. Together with the previous commit, this should free the test
environment from having to set up a configuration.xml file.