ActiveBPELDeployer: new end-to-end test case
Initial test coverage reports showed pretty bad results for the deployers and
the coverage tools. Daniel proposed creating a set of end-to-end tests which
ran on the actual WS-BPEL engines, rather than use the testing deployer which
emulates the WS-BPEL engine through an extra partner track.

This commit adds a new test case with an end-to-end test case which runs in a
running instance of ActiveBPEL. To avoid failing the test suite when ActiveBPEL
is not installed, the test case is automatically skipped if the CATALINA_HOME
environment variable is not set or the $CATALINA_HOME/bpr directory does not
exist.