New XPathVariableResolver for Velocity contexts
To transparently access the variables of the activity context inside
receive conditions, a proper XPath variable resolver has to be provided. This
commit implements that provider.

The provider mostly maps XPath variables to Velocity context variables, except
for one thing. The Xalan engine doesn't map automatically ArrayList objects to
node lists, so we can't simply expect to use something like count($lines) to
count the number of elements in the ArrayList stored in $lines. I tried using
the usual EXSLT functions and even the Xalan extensions to call Java methods,
but I only got "Unknown error in XPath" exceptions one after another.

What did work, however, was to simply detect this case, map the contents of
the ArrayList into a DOM tree and return each element as a node in a node list.
Since org.w3c.dom.NodeList is only an interface, this commit implements a
simple alternative which is backed by an ArrayList (hence ArrayNodeList).

All in all, I like this approach better: the use can simply ignore the problem
above and merrily call "count($lines)" as if it was a regular node list.