PartnerTrack: check assumption to decide whether to skip or not the partner track
Partner tracks (*not* the client track) can now be skipped if the XPath
expression in their assume attribute evaluates to false. This is useful if we
want to use the same test case template as much as possible: some test cases
might require some activity to be run, and some won't need it at all.