ActiveBPEL: throw DeploymentException when the deployment summary shows errors
Currently, BPELUnit only throws DeploymentExceptions when the HTTP response code
from the deployment WS is not a 2XX code. However, ActiveBPEL replies with the 200
response code even for invalid process descriptions which violate some of the
static constraints defined in the WS-BPEL 2.0 standard.

This commit makes BPELUnit parse the deployment summary and throw a DeploymentException
if the deployment summary shows errors (numErrors=X, with X > 0).