TestResult.xsd: relax to match expectations from the code
Just like the BPTS files, test results weren't being validated. Since it's
BPELUnit and not the user writing them, we don't need to validate them as
they are produced, but we should at least have some unit tests for them.

Initial unit tests (to be added in a later commit) revealed that the XML
Schema was stricter than the results actually produced by BPELUnit. This
commit relaxes these constraints so the tests pass.