XMLResultProducerTest: new test case for XMLResultProducer
We need to refactor XMLResultProducerTest so we can reuse the same code for the
inliner, but it's a very important class, so I can't do it without some unit
tests.

This commit adds unit test cases with high overall coverage score for
XMLResultProducer, which check that the results produced conform to the test
result XML Schema. Later commits will add more tests to ensure that the results
are complete as well.