Data sources: get them working with the Eclipse runner
The Eclipse runner does not use the extension registry, but regular extension
points. DataSourceUtil did not take this into account and did not work when the
Eclipse runner was used.

This commit adds the required code to the BPELUnitRunner hierarchy to use the
extension registry for everything except the Eclipse runner and the extension
points when using the Eclipse runner.

It also adds the proper IDs and names to the extension points declared for each
of the data source types, so they are correctly picked up with the right names.

I'm not sure if we should still have the shortName annotations in the data sources.
We pretty much ignore them: the actual short names are those specified in the "type"
attribute in the extension registry and the "id" attribute in the extension points.