org.bpelunit.utils.datasourceinliner: new project
This project will contain a simple utility that "inlines" the data sources in a
test case so an equivalent BPTS is produced, that meets the following constraints:

- the global data source (if any) is moved to the test cases which use it
- each test case template is repeated once per data source row, appending the
  row data at the end of the set up script

This allows for later processing by external tools like IdiginBPEL [1,2],
selecting exactly which test cases should be run to generate the invariants of
the WS-BPEL process. IdiginBPEL can cut and paste the right sections of the .bpts
for regular test cases, but it can't do that reliably for each instance of a test
case template.

[1]: http://idiginbpel.wordpress.com/
[2]: http://gitorious.org/idiginbpel/idiginbpel