BPELUnitBaseRunner: home directory is not needed when using embedded config files
I have revised the standalone distribution to embed the configuration files into
the BPELUnit JAR. This allows the user to simply run BPELUnit and use the (sane)
defaults with no need for the BPELUNIT_HOME environment variable, so long as the
BPTS includes the required deployment information.

In some cases, very little deployment information will be needed: for instance,
the ActiveBPEL deployer assumes by default that the engine is running in a local
instance listening in port 8080. If this assumption is met, the user only needs
to tell BPELUnit the name of the BPR and the type of the deployer ("activebpel").