framework: refactor base URL reading and use lazy initialization for the Jetty local server
The code which read the URL in the BPTS and normalized it was slightly split
between SpecificationLoader and TestSuite. This commit moves it all to
SpecificationLoader and refactors it to a separate method.

The Jetty local server is not created at the same time as TestSuite anymore,
as the user might want to touch up that base URL programmatically before calling
setUp(). This might be useful, for instance, if we want to use a different port
than what is set in the global configuration.