framework: generate OSGi bundle
The basic idea is this:

- Tycho is great for Eclipse projects, but it can't work directly with
  JAR dependencies. Also, if we use it in a project, we can't really
  build regular .jar files that can be placed in a regular Maven repo.

- Felix is great for pure OSGi projects, but it's awkward to work with
  it for the particular case of an Eclipse plugin.

Therefore, the best compromise is to add OSGi bundle artifacts to our
core projects, using Apache Felix, and build the Eclipse-specific parts
with Tycho. However, due to some limitations in Tycho, we cannot mix
Felix and Tycho-based projects in the same reactor. For this reason, the
projects are divided into two reactors. The developer first needs to
build and install all the artifacts in the core reactor, and then the
Tycho reactor can be built.

In the following commits, the rest of the core projects will get OSGi
bundles and the Eclipse projects will be moved to a Tycho-based reactor.