framework: use MANIFEST.MF produced by Apache Felix
The following commits stop using a manually written MANIFEST.MF, and
start using one generated by Apache Felix. The main reason for this is
keeping the Tycho and PDE-based builds as consistent as possible.
Otherwise, we'd use different manifests for them and we wouldn't be
really sure if the released plugins would work or not.