framework: use manual Eclipse files + Maven build
I can't get m-eclipse-p to produce exactly the files that I want, so
we'll have to make do with manually written Eclipse files that take
advantage of the Maven foundation. Basically, they assume "mvn compile"
was run on the main BPELUnit reactor before the whole project is
imported.

Dependencies are grabbed by Maven and placed in the right directory so
the Eclipse project picks them up. Among them, the Eclipse project only
uses those that are really needed. Maven is a little trigger-happy with
the transitive dependencies.