BlackBoard: do not wait for incoming objects if the key won't provide them
I've changed BlackBoard so every key needs to implement the BlackBoardKey
interface. This interface allows keys to say whether they will be able to
provide the requested value or not. If they are unable to, it will be useless
to wait any longer for them, and we should report it as a timeout.

This is mostly useful for partner tracks with 'assume' attributes that will
sometimes run and sometimes not. Even if a matching partner track is found,
if it has been skipped it will be unable to provide any values for the
requests in the blackboard.