toolsupport: try to parse from File, in order to set the XmlBeans SOURCE_NAME document property
Some of the UI may need to access the path where the .bpts file is stored.
For instance, the 'src' attribute of '<template>' elements can store a
relative path from the .bpts to the .vm file used as a Velocity template.

XMLBeans can provide this information through the XmlObject#documentProperties
method, by reading the XmlDocumentProperties.SOURCE_NAME key. However, this
requires that the .bpts document is parsed through a regular Java File object
instead of an input stream.

This commit modifies the BPELUnitEditor class so it will first try to convert
the IFile provided by Eclipse into a File, and parse that. If that doesn't work,
it will revert to the original behaviour of using an input stream.